Move most of state handling to the top container
