+++ nextjs/components/Pages/Channel/index.tsx
@@ -1,4 +1,6 @@
-import PageLayout from '../../layout/PageLayout';
+import React, { useState } from 'react';
+import PageLayout from 'components/layout/PageLayout';
+import { toast } from 'components/Toast';
 import { buildChannelSeo } from 'utilities/seo';
 import { ChannelSerialized } from 'lib/channel';
 import Content from 'components/Pages/Channel/Content/Content';
@@ -6,7 +8,17 @@ import ContentForBots from 'components/Pages/Channel/Content/ContentForBots';
 import { SerializedAccount } from 'serializers/account';
 import { Settings } from 'serializers/account/settings';
 import { SerializedThread } from 'serializers/thread';
+import { SerializedMessage } from 'serializers/message';
 import { Permissions } from 'types/shared';
+import {
+  postReaction,
+  postMerge,
+  moveMessage,
+  moveThread,
+} from './Content/utilities/http';
+import useWebsockets from 'hooks/websockets';
+import useThreadWebsockets from 'hooks/websockets/thread';
+import { ThreadState } from '@prisma/client';
 
 interface Props {
   settings: Settings;
@@ -27,8 +39,8 @@ interface Props {
 }
 
 export default function Channel({
-  threads,
-  pinnedThreads,
+  threads: initialThreads,
+  pinnedThreads: initialPinnedThreads,
   channels,
   currentChannel,
   currentCommunity,
@@ -40,6 +52,378 @@ export default function Channel({
   isBot,
   permissions,
 }: Props) {
+  const [threads, setThreads] = useState<SerializedThread[]>(initialThreads);
+  const [pinnedThreads, setPinnedThreads] =
+    useState<SerializedThread[]>(initialPinnedThreads);
+  const [currentThreadId, setCurrentThreadId] = useState<string>();
+
+  const currentUser = permissions.user || null;
+  const token = permissions.token || null;
+
+  useWebsockets({
+    room: `room:lobby:${currentChannel.id}`,
+    token,
+    permissions,
+    onNewMessage(payload) {
+      try {
+        if (payload.is_reply) {
+          const threadId = payload.thread_id;
+          const messageId = payload.message_id;
+          const imitationId = payload.imitation_id;
+          const message: SerializedMessage =
+            payload.message && JSON.parse(payload.message);
+          if (!message) {
+            return;
+          }
+          setThreads((threads) => {
+            const index = threads.findIndex(({ id }) => id === threadId);
+            const newThreads = [...threads];
+            if (index > -1) {
+              newThreads[index].messages = [
+                ...newThreads[index].messages.filter(
+                  ({ id }) => id !== imitationId && id !== messageId
+                ),
+                message,
+              ];
+            }
+            return newThreads;
+          });
+        }
+
+        if (payload.is_thread) {
+          const threadId = payload.thread_id;
+          const imitationId = payload.imitation_id;
+          const thread: SerializedThread =
+            payload.thread && JSON.parse(payload.thread);
+          if (!thread) {
+            return;
+          }
+          setThreads((threads) => [
+            ...threads.filter(
+              ({ id }) => id !== imitationId && id !== threadId
+            ),
+            thread,
+          ]);
+        }
+      } catch (e) {
+        if (process.env.NODE_ENV === 'development') {
+          console.log(e);
+        }
+      }
+    },
+  });
+
+  useThreadWebsockets({
+    id: currentThreadId,
+    token,
+    permissions,
+    onMessage(message, messageId, imitationId) {
+      setThreads((threads) => {
+        return threads.map((thread) => {
+          if (thread.id === currentThreadId) {
+            return {
+              ...thread,
+              messages: [
+                ...thread.messages.filter(
+                  ({ id }: any) => id !== imitationId && id !== messageId
+                ),
+                message,
+              ],
+            };
+          }
+          return thread;
+        });
+      });
+    },
+  });
+
+  function onSelectThread(thread: SerializedThread) {
+    setCurrentThreadId(thread.id);
+  }
+
+  async function pinThread(threadId: string) {
+    const thread = threads.find(({ id }) => id === threadId);
+    if (!thread) {
+      return;
+    }
+    const newPinned = !thread.pinned;
+    setThreads((threads) => {
+      return threads.map((thread) => {
+        if (thread.id === threadId) {
+          return { ...thread, pinned: newPinned };
+        }
+        return thread;
+      });
+    });
+    setPinnedThreads((pinnedThreads) => {
+      if (newPinned) {
+        return [...pinnedThreads, { ...thread, pinned: true }];
+      } else {
+        return pinnedThreads.filter(({ id }) => id !== threadId);
+      }
+    });
+    return fetch(`/api/threads/${thread.id}`, {
+      method: 'PUT',
+      body: JSON.stringify({ pinned: newPinned }),
+    })
+      .then((response) => {
+        if (response.ok) {
+          return;
+        }
+        throw new Error('Failed to pin the thread.');
+      })
+      .catch((exception) => {
+        alert(exception.message);
+      });
+  }
+
+  async function sendReaction({
+    threadId,
+    messageId,
+    type,
+    active,
+  }: {
+    threadId: string;
+    messageId: string;
+    type: string;
+    active: boolean;
+  }) {
+    function addReaction(threads: SerializedThread[]) {
+      if (!currentUser) {
+        return threads;
+      }
+      return threads.map((thread) => {
+        if (thread.id === threadId) {
+          return {
+            ...thread,
+            messages: thread.messages.map((message) => {
+              if (message.id === messageId) {
+                const reaction = message.reactions.find(
+                  (reaction) => reaction.type === type
+                );
+                if (!reaction) {
+                  return {
+                    ...message,
+                    reactions: [
+                      ...message.reactions,
+                      { type, count: 1, users: [currentUser] },
+                    ],
+                  };
+                }
+
+                if (active) {
+                  return {
+                    ...message,
+                    reactions: message.reactions
+                      .filter((reaction) => {
+                        if (
+                          reaction.type === type &&
+                          reaction.count - 1 === 0
+                        ) {
+                          return false;
+                        }
+                        return true;
+                      })
+                      .map((reaction) => {
+                        if (reaction.type === type) {
+                          const count = reaction.count - 1;
+                          return {
+                            type,
+                            count,
+                            users: reaction.users.filter(
+                              ({ id }) => id !== currentUser.id
+                            ),
+                          };
+                        }
+                        return reaction;
+                      }),
+                  };
+                }
+
+                return {
+                  ...message,
+                  reactions: message.reactions.map((reaction) => {
+                    if (reaction.type === type) {
+                      return {
+                        type,
+                        count: reaction.count + 1,
+                        users: [...reaction.users, currentUser],
+                      };
+                    }
+                    return reaction;
+                  }),
+                };
+              }
+              return message;
+            }),
+          };
+        }
+        return thread;
+      });
+    }
+    setThreads(addReaction);
+    setPinnedThreads(addReaction);
+    postReaction({
+      communityId: currentCommunity?.id,
+      messageId,
+      type,
+      action: active ? 'decrement' : 'increment',
+    });
+  }
+
+  const mergeThreads = ({ from, to }: { from: string; to: string }) => {
+    const source = threads.find((thread) => thread.id === from);
+    const target = threads.find((thread) => thread.id === to);
+
+    setThreads((threads) => {
+      if (!source || !target) {
+        return threads;
+      }
+      return threads
+        .map((thread) => {
+          if (thread.id === from) {
+            return null;
+          }
+          if (thread.id === to) {
+            return {
+              ...thread,
+              messages: [...thread.messages, ...source.messages].sort(
+                (a, b) => {
+                  return (
+                    new Date(a.sentAt).getTime() - new Date(b.sentAt).getTime()
+                  );
+                }
+              ),
+            };
+          }
+          return thread;
+        })
+        .filter(Boolean) as SerializedThread[];
+    });
+    if (!source || !target) {
+      return Promise.resolve();
+    }
+    return postMerge({
+      from: source.id,
+      to: target.id,
+      communityId: currentCommunity?.id,
+    });
+  };
+
+  const moveMessageToThread = ({
+    messageId,
+    threadId,
+  }: {
+    messageId: string;
+    threadId: string;
+  }) => {
+    const messages = [...threads.map((thread) => thread.messages)].flat();
+    const message = messages.find(({ id }) => id === messageId);
+
+    setThreads((threads) => {
+      if (!message) {
+        return threads;
+      }
+      return threads
+        .map((thread) => {
+          if (thread.id === threadId) {
+            const ids = thread.messages.map(({ id }) => id);
+            if (ids.includes(messageId)) {
+              return thread;
+            }
+            return {
+              ...thread,
+              messages: [...thread.messages, message].sort((a, b) => {
+                return (
+                  new Date(a.sentAt).getTime() - new Date(b.sentAt).getTime()
+                );
+              }),
+            };
+          }
+          const ids = thread.messages.map(({ id }) => id);
+          if (ids.includes(messageId)) {
+            return {
+              ...thread,
+              messages: thread.messages.filter(({ id }) => id !== messageId),
+            };
+          }
+          return thread;
+        })
+        .filter(Boolean) as SerializedThread[];
+    });
+
+    return moveMessage({
+      messageId,
+      threadId,
+      communityId: currentCommunity?.id,
+    });
+  };
+
+  const moveThreadToChannel = ({
+    threadId,
+    channelId,
+  }: {
+    threadId: string;
+    channelId: string;
+  }) => {
+    setThreads((threads) => {
+      return threads.filter((thread) => {
+        if (thread.id === threadId && thread.channelId !== channelId) {
+          return false;
+        }
+        return true;
+      });
+    });
+
+    return moveThread({
+      threadId,
+      channelId,
+    });
+  };
+
+  const updateThread = ({
+    state: newState,
+    title: newTitle,
+  }: {
+    state?: ThreadState;
+    title?: string;
+  }) => {
+    if (!currentThreadId) {
+      return;
+    }
+    const options: { state?: ThreadState; title?: string } = {};
+    if (newState) {
+      options.state = newState;
+    }
+    if (newTitle) {
+      options.title = newTitle;
+    }
+    setThreads((threads) => {
+      return threads.map((thread) => {
+        if (thread.id === currentThreadId) {
+          return {
+            ...thread,
+            ...options,
+          };
+        }
+        return thread;
+      });
+    });
+    return fetch(`/api/threads/${currentThreadId}`, {
+      method: 'PUT',
+      body: JSON.stringify(options),
+    })
+      .then((response) => {
+        if (response.ok) {
+          return;
+        }
+        throw new Error('Failed to close the thread.');
+      })
+      .catch((exception) => {
+        toast.error(exception.message);
+      });
+  };
+
   return (
     <PageLayout
       currentChannel={currentChannel}
@@ -69,6 +453,7 @@ export default function Channel({
         />
       ) : (
         <Content
+          key={settings.communityName + channelName}
           threads={threads}
           pinnedThreads={pinnedThreads}
           currentChannel={currentChannel}
@@ -80,7 +465,15 @@ export default function Channel({
           pathCursor={pathCursor}
           isBot={isBot}
           permissions={permissions}
-          key={settings.communityName + channelName}
+          currentThreadId={currentThreadId}
+          setThreads={setThreads}
+          pinThread={pinThread}
+          sendReaction={sendReaction}
+          mergeThreads={mergeThreads}
+          moveMessageToThread={moveMessageToThread}
+          moveThreadToChannel={moveThreadToChannel}
+          onSelectThread={onSelectThread}
+          updateThread={updateThread}
         />
       )}
     </PageLayout>
